barrantes

+++ src/main/java/com/app/usochicamochabackend/review/application/dto/InspectionDTO.java
@@ -12,7 +12,7 @@ public record InspectionDTO(
         String UUID,
         Boolean isUnexpected,
         LocalDateTime dateStamp,
-        Integer hourMeter,
+        Double hourMeter,
         String leakStatus,
         String brakeStatus,
         String beltsPulleysStatus,
