+++ src/main/java/com/app/usochicamochabackend/update/application/dto/PerformChangeHydraulicOilResponse.java
@@ -6,7 +6,7 @@ public record PerformChangeHydraulicOilResponse(
         Long id,
         MachineResponse machine,
         String brand,
-        Integer currentHourMeter,
+        Double currentHourMeter,
         Integer quantity,
         Integer averageHoursChange
 ) {}
